refactor(settings): centralize user-settings shape in shared/settings.ts
The persisted settings object was defined three ways: a full `Settings`
interface in popup.ts and a hand-maintained partial `StoredSettings` in
content.ts (plus inline reads in background.ts). Adding a setting meant
editing both shapes by hand, and forgetting content.ts silently broke the
inline-button / context-menu / PDF flows with no compile error.

Move the shape, defaults, and load/save into src/shared/settings.ts. popup
and content both import it, so the content-side flows now read back exactly
what the popup writes — including the default-merged frontmatter field maps
they previously skipped (a latent consistency fix; default output unchanged).

background.ts and injector.ts keep their bare key reads on purpose: importing
the shared module would pull post-process into the service-worker and injector
bundles just to dedup a string constant. The tweet2md→xclipper storage-key
migration in background.ts is untouched.

No behavior change for default users. tsc clean, 111/111 tests pass, builds.
